requirements.txt
PyQt6>=6.6.0
mss>=9.0.1
Pillow>=10.0.0
numpy>=1.24.0
pynput>=1.7.6
pyinstaller>=6.0.0
